+++ src/schedule/index.ts
@@ -1,1 +1,2 @@
 export { useScheduleStore } from './model';
+export { ScheduleTable, ScheduleTables } from './ui';
