+++ src/schedule/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as ScheduleTable } from './ScheduleTable';
+export { default as ScheduleTables } from './ScheduleTables';
