+++ src/search/index.ts
@@ -1,1 +1,2 @@
 export { useSearchStore } from './model';
+export { SearchDialog, SearchFilter, SearchResult } from './ui';
