+++ src/search/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as SearchDialog } from './SearchDialog';
+export { default as SearchFilter } from './SearchFilter';
+export { default as SearchResult } from './SearchResult';
